socket server, removed db error

+++ services/StorageService.js
@@ -4,12 +4,12 @@ const db = new Sequelize(process.env.DATABASE_URL || urlString, {logging: false}
 const Q = require('q');
 
 const Coin = db.define('coin', {
-	name: {type: db.Sequelize.STRING, unique: true},
-	symbol: {type: db.Sequelize.STRING, unique:true},
-	value: db.Sequelize.INTEGER,
-	tagName: db.Sequelize.STRING,
-	fromAvailable: db.Sequelize.BOOLEAN,
-	toAvailable: db.Sequelize.BOOLEAN
+	name: 				{type: db.Sequelize.STRING, unique: true},
+	symbol: 			{type: db.Sequelize.STRING, unique:true},
+	value: 				db.Sequelize.FLOAT,
+	tagName: 			db.Sequelize.STRING,
+	fromAvailable: 		db.Sequelize.BOOLEAN,
+	toAvailable: 		db.Sequelize.BOOLEAN
 });
 
 const Price = db.define('price', {
